Split settings privacy
