+++ service-worker.js
@@ -189,6 +189,7 @@ const APP_SHELL = [
   '/js/chat-prompt-context.js',
   '/js/chat-summaries.js',
   '/js/settings.js',
+  '/js/settings-privacy.js',
   '/js/settings-data.js',
   '/js/settings-provider-bridge.js',
   '/js/settings-runtime.js',
