+++ tsconfig.checkjs.json
@@ -193,6 +193,7 @@
     "js/recommendations-runtime.js",
     "js/recommendations.js",
     "js/settings.js",
+    "js/settings-privacy.js",
     "js/settings-runtime.js",
     "js/service-worker-update.js",
     "js/shell-actions.js",
