+++ tsconfig.json
@@ -210,6 +210,7 @@
     "js/schema.js",
     "js/schema-environment.js",
     "js/settings.js",
+    "js/settings-privacy.js",
     "js/settings-runtime.js",
     "js/settings-sync-panel.js",
     "js/settings-agent-access-panel.js",
